Resolve "size_t inconsistencies"

Closes #29

See merge request manchester_qbi/manchester_qbi_public/madym_cxx!17

+++ madym/dce_models/mdm_DCEModel2CXM.cxx
@@ -51,10 +51,10 @@ MDM_API std::string mdm_DCEModel2CXM::modelType() const
   return "mdm_DCEModel2CXM";
 }
 
-MDM_API void mdm_DCEModel2CXM::computeCtModel(int nTimes)
+MDM_API void mdm_DCEModel2CXM::computeCtModel(size_t nTimes)
 {
   //Reset all the model concentrations to 0
-  for (int i_t = 0; i_t < nTimes; i_t++)
+  for (size_t i_t = 0; i_t < nTimes; i_t++)
     CtModel_[i_t] = 0;
 
   for (const double& param : pkParams_)
@@ -127,7 +127,7 @@ MDM_API void mdm_DCEModel2CXM::computeCtModel(int nTimes)
   double  Ft_pos = 0;
   double Ft_neg = 0;
 
-  for (int i_t = 1; i_t < nTimes; i_t++)
+  for (size_t i_t = 1; i_t < nTimes; i_t++)
   {
 
     // Get current time, and time change
